+++ packages/react-sdk/CHANGELOG.md
@@ -2,6 +2,20 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.39.3](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-sdk-1.39.2...@stream-io/video-react-sdk-1.39.3) (2026-07-17)
+
+### Dependency Updates
+
+- `@stream-io/audio-filters-web` updated to version `0.10.1`
+- `@stream-io/typescript-config` updated to version `0.1.0`
+- `@stream-io/video-client` updated to version `1.55.2`
+- `@stream-io/video-filters-web` updated to version `0.8.3`
+- `@stream-io/video-react-bindings` updated to version `1.18.3`
+
+### Bug Fixes
+
+- use speakerLayoutSortPreset in RN spotlight view ([#2332](https://github.com/GetStream/stream-video-js/issues/2332)) ([910a5de](https://github.com/GetStream/stream-video-js/commit/910a5de6160cc34de99ea1615ef40bb034c57078))
+
 ## [1.39.2](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-sdk-1.39.1...@stream-io/video-react-sdk-1.39.2) (2026-07-09)
 
 ### Dependency Updates