+++ packages/video-filters-web/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.8.3](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-filters-web-0.8.2...@stream-io/video-filters-web-0.8.3) (2026-07-17)
+
+### Dependency Updates
+
+- `@stream-io/typescript-config` updated to version `0.1.0`
+
+### Bug Fixes
+
+- **video-filters-web:** export BaseVideoProcessor ([#2337](https://github.com/GetStream/stream-video-js/issues/2337)) ([c51b2d6](https://github.com/GetStream/stream-video-js/commit/c51b2d605bc828e3561a909bd929241308dcf4dd))
+
 ## [0.8.2](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-filters-web-0.8.1...@stream-io/video-filters-web-0.8.2) (2026-07-09)
 
 ### Dependency Updates